docker/srcs/uwsgi-django/requirements.txt
# docker/srcs/uwsgi-django/requirements.txt
# このファイルは、プロジェクトに必要なPythonパッケージとそのバージョンを記述します。

# フレームワーク
Django==5.0.2

# PostgreSQLデータベース用
psycopg2==2.9.9

# WSGIサーバー
uwsgi>=2.0

django-prometheus==2.3.1
python-logstash==0.4.8
web3==6.17.0
python-dotenv==1.0.1

Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
myst-parser==2.0.0

requests==2.31.0
parsimonious==0.10.0

# 2FA
django-otp==1.3.0
qrcode==7.4.2
pyotp==2.9.0

# JWT
djangorestframework-simplejwt==5.3.1

# chat
channels==4.1.0
daphne==4.1.2
django-shortuuidfield==0.1.3
